+++ features/auth/src/main/res/layout/activity_sign_in.xml
@@ -8,12 +8,17 @@
         android:layout_height="match_parent"
         tools:context=".SignInActivity">
 
-        <TextView
-            android:id="@+id/text_view"
+        <androidx.fragment.app.FragmentContainerView
+            android:id="@+id/nav_hos_fragment"
+            android:name="androidx.navigation.fragment.NavHostFragment"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:gravity="center"
-            android:text="@string/title_sign_in"
-            android:textSize="32sp" />
+            app:defaultNavHost="true"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent"
+            app:navGraph="@navigation/navigation_sign_in" />
+
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
